Struct multimap::OccupiedEntry

pub struct OccupiedEntry<'a, K: 'a, V: 'a> { /* fields omitted */ }

A view into a single occupied location in a MultiMap.

Methods

impl<'a, K: 'a, V: 'a> OccupiedEntry<'a, K, V>

fn get(&self) -> &V

Gets a reference to the first item in value in the vector corresponding to entry.

fn get_vec(&self) -> &Vec<V>

Gets a reference to the values (vector) corresponding to entry.

fn get_mut(&mut self) -> &mut V

Gets a mut reference to the first item in value in the vector corresponding to entry.

fn get_vec_mut(&mut self) -> &mut Vec<V>

Gets a mut reference to the values (vector) corresponding to entry.

fn into_mut(self) -> &'a mut V

Converts the OccupiedEntry into a mutable reference to the first item in value in the entry with a lifetime bound to the map itself

fn into_vec_mut(self) -> &'a mut Vec<V>

Converts the OccupiedEntry into a mutable reference to the values (vector) in the entry with a lifetime bound to the map itself

fn insert(&mut self, value: V)

Inserts a new value onto the vector of the entry.

fn insert_vec(&mut self, values: Vec<V>)

Extends the existing vector with the specified values.

fn remove(self) -> Vec<V>

Takes the values (vector) out of the entry, and returns it