Struct multimap::MultiMap

pub struct MultiMap<K, V, S = RandomState> { /* private fields */ }

Implementations

impl<K, V> MultiMap<K, V> where
K: Eq + Hash,

pub fn new() -> MultiMap<K, V>

Creates an empty MultiMap

Examples

use multimap::MultiMap;

let mut map: MultiMap<&str, isize> = MultiMap::new();

pub fn with_capacity(capacity: usize) -> MultiMap<K, V>

Creates an empty multimap with the given initial capacity.

Examples

use multimap::MultiMap;

let mut map: MultiMap<&str, isize> = MultiMap::with_capacity(20);

impl<K, V, S> MultiMap<K, V, S> where
K: Eq + Hash,
S: BuildHasher,

pub fn with_hasher(hash_builder: S) -> MultiMap<K, V, S>

Creates an empty MultiMap which will use the given hash builder to hash keys.

Examples

use multimap::MultiMap;
use std::collections::hash_map::RandomState;

let s = RandomState::new();
let mut map: MultiMap<&str, isize> = MultiMap::with_hasher(s);

pub fn with_capacity_and_hasher(
capacity: usize,
hash_builder: S
) -> MultiMap<K, V, S>

Creates an empty MultiMap with the given intial capacity and hash builder.

Examples

use multimap::MultiMap;
use std::collections::hash_map::RandomState;

let s = RandomState::new();
let mut map: MultiMap<&str, isize> = MultiMap::with_capacity_and_hasher(20, s);

pub fn insert(&mut self, k: K, v: V)

Inserts a key-value pair into the multimap. If the key does exist in the map then the value is pushed to that key’s vector. If the key doesn’t exist in the map a new vector with the given value is inserted.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert("key", 42);

pub fn insert_many<I: IntoIterator<Item = V>>(&mut self, k: K, v: I)

Inserts multiple key-value pairs into the multimap. If the key does exist in the map then the values are extended into that key’s vector. If the key doesn’t exist in the map a new vector collected from the given values is inserted.

This may be more efficient than inserting values independently.

Examples

use multimap::MultiMap;

let mut map = MultiMap::<&str, &usize>::new();
map.insert_many("key", &[42, 43]);

pub fn insert_many_from_slice(&mut self, k: K, v: &[V]) where
V: Clone,

Inserts multiple key-value pairs into the multimap. If the key does exist in the map then the values are extended into that key’s vector. If the key doesn’t exist in the map a new vector collected from the given values is inserted.

This may be more efficient than inserting values independently.

Examples

use multimap::MultiMap;

let mut map = MultiMap::<&str, usize>::new();
map.insert_many_from_slice("key", &[42, 43]);

pub fn contains_key<Q: ?Sized>(&self, k: &Q) -> bool where
K: Borrow<Q>,
Q: Eq + Hash,

Returns true if the map contains a value for the specified key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1, 42);
assert_eq!(map.contains_key(&1), true);
assert_eq!(map.contains_key(&2), false);

pub fn len(&self) -> usize

Returns the number of elements in the map.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1, 42);
map.insert(2, 1337);
assert_eq!(map.len(), 2);

pub fn remove<Q: ?Sized>(&mut self, k: &Q) -> Option<Vec<V>> where
K: Borrow<Q>,
Q: Eq + Hash,

Removes a key from the map, returning the vector of values at the key if the key was previously in the map.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1, 42);
map.insert(1, 1337);
assert_eq!(map.remove(&1), Some(vec![42, 1337]));
assert_eq!(map.remove(&1), None);

pub fn get<Q: ?Sized>(&self, k: &Q) -> Option<&V> where
K: Borrow<Q>,
Q: Eq + Hash,

Returns a reference to the first item in the vector corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1, 42);
map.insert(1, 1337);
assert_eq!(map.get(&1), Some(&42));

pub fn get_mut<Q: ?Sized>(&mut self, k: &Q) -> Option<&mut V> where
K: Borrow<Q>,
Q: Eq + Hash,

Returns a mutable reference to the first item in the vector corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1, 42);
map.insert(1, 1337);
if let Some(v) = map.get_mut(&1) {
    *v = 99;
}
assert_eq!(map[&1], 99);

pub fn get_vec<Q: ?Sized>(&self, k: &Q) -> Option<&Vec<V>> where
K: Borrow<Q>,
Q: Eq + Hash,

Returns a reference to the vector corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1, 42);
map.insert(1, 1337);
assert_eq!(map.get_vec(&1), Some(&vec![42, 1337]));

pub fn get_vec_mut<Q: ?Sized>(&mut self, k: &Q) -> Option<&mut Vec<V>> where
K: Borrow<Q>,
Q: Eq + Hash,

Returns a mutable reference to the vector corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1, 42);
map.insert(1, 1337);
if let Some(v) = map.get_vec_mut(&1) {
    (*v)[0] = 1991;
    (*v)[1] = 2332;
}
assert_eq!(map.get_vec(&1), Some(&vec![1991, 2332]));

pub fn is_vec<Q: ?Sized>(&self, k: &Q) -> bool where
K: Borrow<Q>,
Q: Eq + Hash,

Returns true if the key is multi-valued.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1, 42);
map.insert(1, 1337);
map.insert(2, 2332);

assert_eq!(map.is_vec(&1), true);   // key is multi-valued
assert_eq!(map.is_vec(&2), false);  // key is single-valued
assert_eq!(map.is_vec(&3), false);  // key not in map

pub fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

Examples

use multimap::MultiMap;

let map: MultiMap<usize, usize> = MultiMap::new();
assert!(map.capacity() >= 0);

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
assert!(map.is_empty());
map.insert(1,42);
assert!(!map.is_empty());

pub fn clear(&mut self)

Clears the map, removing all key-value pairs. Keeps the allocated memory for reuse.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1,42);
map.clear();
assert!(map.is_empty());

pub fn keys<'a>(&'a self) -> Keys<'a, K, Vec<V>>

An iterator visiting all keys in arbitrary order. Iterator element type is &’a K.

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1,42);
map.insert(2,1337);
map.insert(4,1991);

for key in map.keys() {
    println!("{:?}", key);
}

pub fn iter(&self) -> Iter<'_, K, V>

An iterator visiting all key-value pairs in arbitrary order. The iterator returns a reference to the key and the first element in the corresponding key’s vector. Iterator element type is (&’a K, &’a V).

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1,42);
map.insert(1,1337);
map.insert(3,2332);
map.insert(4,1991);

for (key, value) in map.iter() {
    println!("key: {:?}, val: {:?}", key, value);
}

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

An iterator visiting all key-value pairs in arbitrary order. The iterator returns a reference to the key and a mutable reference to the first element in the corresponding key’s vector. Iterator element type is (&’a K, &’a mut V).

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1,42);
map.insert(1,1337);
map.insert(3,2332);
map.insert(4,1991);

for (_, value) in map.iter_mut() {
    *value *= *value;
}

for (key, value) in map.iter() {
    println!("key: {:?}, val: {:?}", key, value);
}

pub fn iter_all(&self) -> IterAll<'_, K, Vec<V>>

An iterator visiting all key-value pairs in arbitrary order. The iterator returns a reference to the key and the corresponding key’s vector. Iterator element type is (&’a K, &’a V).

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1,42);
map.insert(1,1337);
map.insert(3,2332);
map.insert(4,1991);

for (key, values) in map.iter_all() {
    println!("key: {:?}, values: {:?}", key, values);
}

pub fn iter_all_mut(&mut self) -> IterAllMut<'_, K, Vec<V>>

An iterator visiting all key-value pairs in arbitrary order. The iterator returns a reference to the key and the corresponding key’s vector. Iterator element type is (&’a K, &’a V).

Examples

use multimap::MultiMap;

let mut map = MultiMap::new();
map.insert(1,42);
map.insert(1,1337);
map.insert(3,2332);
map.insert(4,1991);

for (key, values) in map.iter_all_mut() {
    for value in values.iter_mut() {
        *value = 99;
    }
}

for (key, values) in map.iter_all() {
    println!("key: {:?}, values: {:?}", key, values);
}

pub fn entry(&mut self, k: K) -> Entry<'_, K, V>

Gets the specified key’s corresponding entry in the map for in-place manipulation. It’s possible to both manipulate the vector and the ‘value’ (the first value in the vector).

Examples

use multimap::MultiMap;

let mut m = MultiMap::new();
m.insert(1, 42);

{
    let mut v = m.entry(1).or_insert(43);
    assert_eq!(v, &42);
    *v = 44;
}
assert_eq!(m.entry(2).or_insert(666), &666);

{
    let mut v = m.entry(1).or_insert_vec(vec![43]);
    assert_eq!(v, &vec![44]);
    v.push(50);
}
assert_eq!(m.entry(2).or_insert_vec(vec![666]), &vec![666]);

assert_eq!(m.get_vec(&1), Some(&vec![44, 50]));

pub fn retain<F>(&mut self, f: F) where
F: FnMut(&K, &V) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all pairs (k, v) such that f(&k,&mut v) returns false.

Examples

use multimap::MultiMap;

let mut m = MultiMap::new();
m.insert(1, 42);
m.insert(1, 99);
m.insert(2, 42);
m.retain(|&k, &v| { k == 1 && v == 42 });
assert_eq!(1, m.len());
assert_eq!(Some(&42), m.get(&1));