Struct multimap::OccupiedEntry

pub struct OccupiedEntry<'a, K: 'a, V: 'a> { /* fields omitted */ }

A view into a single occupied location in a MultiMap.

Methods

impl<'a, K: 'a, V: 'a> OccupiedEntry<'a, K, V>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get(&self) -> &V

Gets a reference to the first item in value in the vector corresponding to entry.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_vec(&self) -> &Vec<V>

Gets a reference to the values (vector) corresponding to entry.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_mut(&mut self) -> &mut V

Gets a mut reference to the first item in value in the vector corresponding to entry.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_vec_mut(&mut self) -> &mut Vec<V>

Gets a mut reference to the values (vector) corresponding to entry.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn into_mut(self) -> &'a mut V

Converts the OccupiedEntry into a mutable reference to the first item in value in the entry with a lifetime bound to the map itself

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn into_vec_mut(self) -> &'a mut Vec<V>

Converts the OccupiedEntry into a mutable reference to the values (vector) in the entry with a lifetime bound to the map itself

pub fn insert(&mut self, value: V)

Inserts a new value onto the vector of the entry.

pub fn insert_vec(&mut self, values: Vec<V>)

Extends the existing vector with the specified values.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn remove(self) -> Vec<V>

Takes the values (vector) out of the entry, and returns it